third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGTextElement-dom-rotate-attr-expected.txt

SVG 1.1 dynamic update tests
Text content

Tests dynamic updates of the 'rotate' attribute of the SVGTextElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textElement.getAttribute('rotate') is "0"
PASS textElement.getAttribute('rotate') is "20"
PASS successfullyParsed is true

TEST COMPLETE
